add Skaffold
With skaffold we dont have to update image,push to hub and restart deployment. We spesify the code skaffold should watch for changes, and when it catches them skaffold  will update our POD. If the changes we spesified are different than the ones skaffold caught then skaffold  will push to hub and restart deployment on its own.